backend/requirements.txt
fastapi
uvicorn
joblib
numpy
pandas
scikit-learn
python-whois
requests
beautifulsoup4
python-multipart
jinja2
psycopg2-binary
bcrypt
itsdangerous